Added methods to and annotations

+++ src/main/java/app/persistence/IDAO.java
@@ -5,11 +5,11 @@ import java.util.Set;
 public interface IDAO <T> {
     T create(T t);
 
-    T getById(int id);
+    T getById(long id);
 
     T update(T t);
 
-    T delete(int id);
+    T delete(long id);
 
     Set<T> getAll();
 }
